Refactor ONNXShapeHelper code.
