+++ src/Conversion/ONNXToKrnl/Tensor/ArgMax.cpp
@@ -6,7 +6,7 @@
 
 #include "mlir/Dialect/Affine/IR/AffineOps.h"
 #include "src/Conversion/ONNXToKrnl/ONNXToKrnlCommon.hpp"
-#include "src/Dialect/ONNX/ONNXShapeHelper.hpp"
+#include "src/Dialect/ONNX/ShapeInference/ONNXShapeHelper.hpp"
 
 using namespace mlir;
