+++ src/Dialect/ONNX/ShapeInference/Slice.cpp
@@ -1,0 +1,146 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+//===------------- Slice.cpp - Shape Inference for Slice Op ---------------===//
+//
+// This file implements shape inference for the ONNX Slice Operator.
+//
+//===----------------------------------------------------------------------===//
+
+#include "src/Dialect/ONNX/ShapeInference/ONNXShapeHelper.hpp"
+
+ONNXSliceOpShapeHelper::ONNXSliceOpShapeHelper(ONNXSliceOp *newOp)
+    : ONNXOpShapeHelper<ONNXSliceOp>(
+          newOp, newOp->getOperation()->getNumResults()),
+      starts(), ends(), steps() {}
+
+ONNXSliceOpShapeHelper::ONNXSliceOpShapeHelper(ONNXSliceOp *newOp,
+    OpBuilder *rewriter, ArrayValueIndexCapture::GetDenseVal fGetDenseVal,
+    ArrayValueIndexCapture::LoadVal fLoadVal)
+    : ONNXOpShapeHelper<ONNXSliceOp>(newOp,
+          newOp->getOperation()->getNumResults(), rewriter, fGetDenseVal,
+          fLoadVal),
+      starts(), ends(), steps() {}
+
+LogicalResult ONNXSliceOpShapeHelper::computeShape(
+    ONNXSliceOpAdaptor operandAdaptor) {
+  // Shape inference indicated by passing a null rewriter pointer.
+  // Output dims of results.
+  DimsExpr outputDims;
+
+  // Get info about input data operand.
+  Value data = operandAdaptor.data();
+  uint64_t dataRank = data.getType().cast<ShapedType>().getShape().size();
+
+  // Get each of the axes, and save the literal values in axesIntLit.
+  SmallVector<int64_t, 2> axesIntLit;
+  Value axes = operandAdaptor.axes();
+  if (axes.getType().isa<NoneType>()) {
+    // If `axes` are omitted, they are set to `[0, ..., nDim-1]`."
+    for (unsigned int i = 0; i < dataRank; ++i)
+      axesIntLit.emplace_back(i);
+  } else if (auto valueAttribute = fGetDenseVal(axes)) {
+    // If `axes` are constants, read them."
+    for (IntegerAttr value : valueAttribute.getValues<IntegerAttr>()) {
+      int64_t axis = value.cast<IntegerAttr>().getInt();
+      if (axis < 0)
+        axis += dataRank;
+      if (!(axis >= 0 && axis < (int64_t)dataRank))
+        return op->emitError("Axes contains an out-of-bound index");
+      axesIntLit.emplace_back(axis);
+    }
+  } else {
+    return op->emitError("Axes must be known at compile time");
+  }
+  uint64_t sliceRank = axesIntLit.size();
+
+  // Initialize context and results (start & output)
+  starts.resize(dataRank);
+  steps.resize(dataRank);
+  ends.resize(dataRank);
+  outputDims.resize(dataRank);
+
+  // SmallVector<uint64_t, 1> index1D(1, 0);
+  ArrayValueIndexCapture startsCapture(
+      operandAdaptor.starts(), fGetDenseVal, fLoadVal);
+  ArrayValueIndexCapture endsCapture(
+      operandAdaptor.ends(), fGetDenseVal, fLoadVal);
+  ArrayValueIndexCapture stepsCapture(
+      operandAdaptor.steps(), fGetDenseVal, fLoadVal);
+  MemRefBoundsIndexCapture dataBounds(data);
+  for (uint64_t i = 0; i < sliceRank; i++) {
+    // i is index in start/step/end/output
+    // ii is logical index in mem/loop bounds
+    int ii = axesIntLit[i];
+    // Get start, end, step, and dim index expressions.
+    // Get start.
+    SymbolIndexExpr startInput(startsCapture.getSymbol(i));
+    if (startInput.isUndefined())
+      return op->emitError("start input parameter could not be processed");
+    // Get end.
+    SymbolIndexExpr endInput(endsCapture.getSymbol(i));
+    if (endInput.isUndefined())
+      return op->emitError("end input parameter could not be processed");
+    // Get step.
+    SymbolIndexExpr stepInput(stepsCapture.getSymbol(i));
+    if (stepInput.isUndefined())
+      return op->emitError("step input parameter could not be processed");
+    if (stepInput.isLiteral() && stepInput.getLiteral() == 0)
+      return op->emitError("step input parameter cannot be zero");
+    // Get dim.
+    DimIndexExpr dimInput(dataBounds.getDim(ii));
+
+    // Now proceed with the computations for start/end/dim.
+    // Calculation for start: start < 0 ? start + dim : start.
+    IndexExpr startPos =
+        IndexExpr::select(startInput < 0, startInput + dimInput, startInput);
+    // Step < 0: clamp(0, start, dim -1) else clamp(0, start, dim)
+    IndexExpr neg = startPos.clamp(0, dimInput - 1);
+    IndexExpr pos = startPos.clamp(0, dimInput);
+    IndexExpr startFinal = IndexExpr::select(stepInput < 0, neg, pos);
+
+    // Calculation for end: end<0 -> end + dim else -> end;
+    // special case end <= -inf -> -1;  end >= inf -> dim;
+    int64_t negInf = std::numeric_limits<int32_t>::min();
+    int64_t posInf = std::numeric_limits<int32_t>::max();
+    IndexExpr endPos =
+        IndexExpr::select(endInput < 0, endInput + dimInput, endInput);
+    endPos = endPos.selectOrSelf(endInput <= negInf, -1);
+    endPos = endPos.selectOrSelf(endInput >= posInf, dimInput);
+    // End: step<0: clamp(-1, end, dim); step>0 clamp(0, end, dim)
+    neg = endPos.clamp(-1, dimInput);
+    pos = endPos.clamp(0, dimInput);
+    IndexExpr endFinal = IndexExpr::select(stepInput < 0, neg, pos);
+
+    // Calculation for output size.
+    IndexExpr dimOutputFinal = (endFinal - startFinal).ceilDiv(stepInput);
+    // should use a max
+    dimOutputFinal = dimOutputFinal.selectOrSelf(dimOutputFinal < 0, 0);
+
+    // Save results
+    starts[ii] = startFinal;
+    steps[ii] = stepInput;
+    ends[ii] = endFinal;
+    outputDims[ii] = dimOutputFinal;
+  }
+
+  // Handle the default for the non-axis arrays; they are detected with 0
+  // steps (illegal value).
+  for (uint64_t i = 0; i < dataRank; ++i) {
+    if (steps[i].isUndefined()) {
+      // have one unset, put the defaults (start was already at zero, so we
+      // are fine).
+      starts[i] = LiteralIndexExpr(0);
+      steps[i] = LiteralIndexExpr(1);
+      DimIndexExpr dimInput(dataBounds.getDim(i));
+      ends[i] = dimInput;
+      outputDims[i] = dimInput;
+    }
+  }
+
+  // Save the final result.
+  dimsForOutput(0) = outputDims;
+
+  return success();
+}
